website/backend/requirements.txt
# RepoPilot Website — Lightweight Backend Dependencies
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Git operations
gitpython>=3.1.41

# Gemini LLM (architecture analysis)
google-generativeai>=0.8.0

# Language detection
pygments>=2.17.2

# HTTP client
httpx>=0.26.0

# Logging
structlog>=24.1.0
